fix: improved the landing page's modular components
